+++ UserDev/ClusterStudy/CMAlgoOutOfConeSeparate.cc
@@ -35,8 +35,8 @@ namespace cluster {
     double len1 = cluster1.GetParams().length;
     double len2 = cluster2.GetParams().length;
     
-    int hits1 = cluster1.GetHitVector().size();
-    int hits2 = cluster2.GetHitVector().size();
+    size_t hits1 = cluster1.GetHitVector().size();
+    size_t hits2 = cluster2.GetHitVector().size();
 
     double startseparation = (w_start2-w_start1)*(w_start2-w_start1) + (t_start2-t_start1)*(t_start2-t_start1);
     //convert sepration to be instead of just angle -> angle/distance^n (n=1 for now)
